refactor: re-arrange package structure
