+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -1,8 +1,7 @@
-package step2
+package calculator
 
 import java.util.LinkedList
 import java.util.Queue
-import java.util.function.BiFunction
 
 private const val BASIC_SPLITTER = " "
 
@@ -58,27 +57,3 @@ private fun String.toOperatorQueue(splitter: String): Queue<Operator> {
 private fun Int.isNumberIndex(): Boolean = this % 2 == 0
 
 private fun Int.isOperatorIndex(): Boolean = this % 2 != 0
-
-enum class Operator(private val operation: BiFunction<Double, Double, Double>) {
-    PLUS({ a, b -> a + b }),
-    MINUS({ a, b -> a - b }),
-    MULTIPLY({ a, b -> a * b }),
-    DIVIDE({ a, b -> a / b })
-    ;
-
-    fun calculate(a: Double, b: Double): Double {
-        return operation.apply(a, b)
-    }
-
-    companion object {
-        fun findBy(symbol: String): Operator {
-            return when (symbol) {
-                "+" -> PLUS
-                "-" -> MINUS
-                "*" -> MULTIPLY
-                "/" -> DIVIDE
-                else -> throw IllegalArgumentException("지원하지 않는 연산자입니다. symbol: $symbol")
-            }
-        }
-    }
-}
