+++ src/test/kotlin/calculator/StringCalculatorTest.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
@@ -9,30 +9,6 @@ import org.junit.jupiter.params.provider.ValueSource
 
 class StringCalculatorTest {
     private val calculator = StringCalculator()
-
-    @Test
-    fun 더하기() {
-        val result = Operator.PLUS.calculate(1.0, 2.0)
-        assertThat(result).isEqualTo(3.0)
-    }
-
-    @Test
-    fun 곱하기() {
-        val result = Operator.MULTIPLY.calculate(1.0, 2.0)
-        assertThat(result).isEqualTo(2.0)
-    }
-
-    @Test
-    fun 빼기() {
-        val result = Operator.MINUS.calculate(1.0, 2.0)
-        assertThat(result).isEqualTo(-1.0)
-    }
-
-    @Test
-    fun 나누기() {
-        val result = Operator.DIVIDE.calculate(1.0, 2.0)
-        assertThat(result).isEqualTo(0.5)
-    }
 
     @ParameterizedTest
     @CsvSource(value = ["1,2,+,3", "1,2,*,2", "1,2,-,-1", "1,2,/,0.5"])
