+++ src/test/kotlin/study/PersonKoTest.kt
@@ -1,4 +1,4 @@
-package step1
+package study
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
